Gestion métier Stagiaire avec renforcement sécurité accès modif profil + gestion sécu authorisation ajout piece jointe

+++ src/main/java/com/hrizzon2/demotest/service/stagiaire/StagiaireService.java
@@ -38,5 +38,5 @@ public interface StagiaireService {
     Inscription inscrireStagiaire(Stagiaire stagiaire, Formation formation);
 
 
-    Long getIdFromPrincipal(Principal principal);
+    Integer getIdFromPrincipal(Principal principal);
 }
